Add files via upload
This is a simple web-server coded with C language. It is working using TCP/IP protocol.
You can bind it on any available ip/port. 

To start up you just need to unpack "webroot.zip" to "root/Desktop"(Here is index.html and a simple picture) and run SwS.out(Simple Web Server).
So... It is quite simple and has many vulnerabilities and you can train your hacking skills on it. For example this server has buffer overflow vulnerability and you can try to overwrite the return address of the function and inject the shellcode. Good luck!